Create subcomponents for ObsEditHeader and AddEvidenceModal
