+++ src/components/ObsEdit/ObsEdit.js
@@ -1,20 +1,11 @@
 // @flow
 
-import {
-  BottomSheetBackdrop,
-  BottomSheetModal,
-  BottomSheetModalProvider
-} from "@gorhom/bottom-sheet";
-import { HeaderBackButton } from "@react-navigation/elements";
 import { useFocusEffect, useNavigation, useRoute } from "@react-navigation/native";
-import { MAX_PHOTOS_ALLOWED } from "components/Camera/StandardCamera";
 import MediaViewer from "components/MediaViewer/MediaViewer";
 import MediaViewerModal from "components/MediaViewer/MediaViewerModal";
 import Button from "components/SharedComponents/Buttons/Button";
-import EvidenceButton from "components/SharedComponents/Buttons/EvidenceButton";
-import KebabMenu from "components/SharedComponents/KebabMenu";
 import ScrollNoFooter from "components/SharedComponents/ScrollNoFooter";
-import { Pressable, Text, View } from "components/styledComponents";
+import { Text, View } from "components/styledComponents";
 import { ObsEditContext, RealmContext } from "providers/contexts";
 import type { Node } from "react";
 import React, {
@@ -23,18 +14,16 @@ import React, {
 } from "react";
 import { useTranslation } from "react-i18next";
 import { BackHandler } from "react-native";
-import { Menu } from "react-native-paper";
-import Icon from "react-native-vector-icons/MaterialIcons";
 import Photo from "realmModels/Photo";
 import fetchUserLocation from "sharedHelpers/fetchUserLocation";
 import useLocalObservation from "sharedHooks/useLocalObservation";
 import useLoggedIn from "sharedHooks/useLoggedIn";
-import { textStyles, viewStyles } from "styles/obsEdit/obsEdit";
-import colors from "styles/tailwindColors";
+import { viewStyles } from "styles/obsEdit/obsEdit";
 
-import DeleteObservationDialog from "./DeleteObservationDialog";
+import AddEvidenceModal from "./AddEvidenceModal";
 import EvidenceSection from "./EvidenceSection";
 import IdentificationSection from "./IdentificationSection";
+import ObsEditHeader from "./ObsEditHeader";
 import OtherDataSection from "./OtherDataSection";
 
 const { useRealm } = RealmContext;
@@ -45,7 +34,6 @@ const TARGET_POSITIONAL_ACCURACY = 10;
 const ObsEdit = ( ): Node => {
   const {
     currentObsIndex,
-    setCurrentObsIndex,
     observations,
     saveObservation,
     saveAndUploadObservation,
@@ -53,10 +41,13 @@ const ObsEdit = ( ): Node => {
     updateObservationKeys
   } = useContext( ObsEditContext );
   const currentObs = observations[currentObsIndex];
+  const obsPhotos = currentObs?.observationPhotos;
+  const photoUris = obsPhotos && Array.from( obsPhotos ).map(
+    obsPhoto => Photo.displayLocalOrRemoteSquarePhoto( obsPhoto.photo )
+  );
   const navigation = useNavigation( );
   const { params } = useRoute( );
   const { t } = useTranslation( );
-  const bottomSheetModalRef = useRef( null );
   const localObservation = useLocalObservation( params?.uuid );
 
   useEffect( ( ) => {
@@ -72,24 +63,14 @@ const ObsEdit = ( ): Node => {
   const isLoggedIn = useLoggedIn( );
   const [mediaViewerVisible, setMediaViewerVisible] = useState( false );
   const [initialPhotoSelected, setInitialPhotoSelected] = useState( null );
-  const [photoUris, setPhotoUris] = useState( [] );
-  const [snapPoint, setSnapPoint] = useState( 150 );
-  const [deleteDialogVisible, setDeleteDialogVisible] = useState( false );
   const [shouldFetchLocation, setShouldFetchLocation] = useState( !currentObs?._created_at );
   const [fetchingLocation, setFetchingLocation] = useState( false );
   const [positionalAccuracy, setPositionalAccuracy] = useState( INITIAL_POSITIONAL_ACCURACY );
   const mountedRef = useRef( true );
-
-  const disableAddingMoreEvidence = photoUris.length >= MAX_PHOTOS_ALLOWED;
+  const [showAddEvidenceModal, setShowAddEvidenceModal] = useState( false );
 
   const showModal = ( ) => setMediaViewerVisible( true );
   const hideModal = ( ) => setMediaViewerVisible( false );
-
-  const showNextObservation = ( ) => setCurrentObsIndex( currentObsIndex + 1 );
-  const showPrevObservation = ( ) => setCurrentObsIndex( currentObsIndex - 1 );
-
-  const showDialog = ( ) => setDeleteDialogVisible( true );
-  const hideDialog = ( ) => setDeleteDialogVisible( false );
 
   const handleBackButtonPress = useCallback( async ( ) => {
     setObservations( [] );
@@ -114,42 +95,6 @@ const ObsEdit = ( ): Node => {
 
       return ( ) => BackHandler.removeEventListener( "hardwareBackPress", onBackPress );
     }, [handleBackButtonPress] )
-  );
-
-  const renderKebabMenu = ( ) => (
-    <>
-      <DeleteObservationDialog
-        deleteDialogVisible={deleteDialogVisible}
-        hideDialog={hideDialog}
-      />
-      <KebabMenu>
-        <Menu.Item
-          onPress={showDialog}
-          title={t( "Delete" )}
-        />
-      </KebabMenu>
-    </>
-  );
-
-  const renderHeader = ( ) => (
-    <View className="flex-row justify-between">
-      <HeaderBackButton onPress={handleBackButtonPress} tintColor={colors.black} />
-      {observations.length === 1
-        ? <Text className="text-2xl">{t( "New-Observation" )}</Text>
-        : (
-          <View className="flex-row items-center">
-            <Pressable onPress={showPrevObservation} className="w-16">
-              {currentObsIndex !== 0 && <Icon name="keyboard-arrow-left" size={30} />}
-            </Pressable>
-            <Text className="text-2xl">{`${currentObsIndex + 1} of ${observations.length}`}</Text>
-            <Pressable onPress={showNextObservation} className="w-16">
-              {( currentObsIndex !== observations.length - 1 )
-                && <Icon name="keyboard-arrow-right" size={30} />}
-            </Pressable>
-          </View>
-        )}
-      {renderKebabMenu( )}
-    </View>
   );
 
   // Hook version of componentWillUnmount. We use a ref to track mounted
@@ -247,49 +192,12 @@ const ObsEdit = ( ): Node => {
     showModal( );
   };
 
-  useEffect( ( ) => {
-    if ( !currentObs || !currentObs.observationPhotos ) { return; }
-    const uris = Array.from( currentObs.observationPhotos ).map(
-      obsPhoto => Photo.displayLocalOrRemoteSquarePhoto( obsPhoto.photo )
-    );
-    setPhotoUris( uris );
-  }, [currentObs] );
-
-  const addEvidence = () => {
-    bottomSheetModalRef.current?.present();
-  };
+  const addEvidence = ( ) => setShowAddEvidenceModal( true );
 
   if ( !currentObs ) { return null; }
 
-  const renderBackdrop = props => (
-    <BottomSheetBackdrop
-      // eslint-disable-next-line react/jsx-props-no-spreading
-      {...props}
-      pressBehavior="close"
-      appearsOnIndex={0}
-      disappearsOnIndex={-1}
-    />
-  );
-
-  const onImportPhoto = async () => {
-    navigation.navigate( "PhotoGallery", { photos: photoUris, skipGroupPhotos: true } );
-
-    bottomSheetModalRef.current?.dismiss();
-  };
-
-  const onTakePhoto = async () => {
-    navigation.navigate( "StandardCamera", { photos: photoUris } );
-
-    bottomSheetModalRef.current?.dismiss();
-  };
-
-  const onRecordSound = () => {
-    // TODO - need to implement
-    console.log( "Record sound" );
-  };
-
   return (
-    <BottomSheetModalProvider>
+    <>
       <MediaViewerModal
         mediaViewerVisible={mediaViewerVisible}
         hideModal={hideModal}
@@ -302,7 +210,7 @@ const ObsEdit = ( ): Node => {
         />
       </MediaViewerModal>
       <ScrollNoFooter style={mediaViewerVisible && viewStyles.mediaViewerSafeAreaView}>
-        {renderHeader( )}
+        <ObsEditHeader handleBackButtonPress={handleBackButtonPress} />
         <Text className="text-2xl ml-4">{t( "Evidence" )}</Text>
         <EvidenceSection
           handleSelection={handleSelection}
@@ -329,59 +237,12 @@ const ObsEdit = ( ): Node => {
             disabled={!isLoggedIn}
           />
         </View>
-
-        <BottomSheetModal
-          ref={bottomSheetModalRef}
-          index={0}
-          enableOverDrag={false}
-          enablePanDownToClose={false}
-          snapPoints={[snapPoint]}
-          backdropComponent={renderBackdrop}
-        >
-          <View
-            className="items-center p-10"
-            onLayout={( {
-              nativeEvent: {
-                layout: { height }
-              }
-            } ) => {
-              setSnapPoint( height + 50 );
-            }}
-          >
-            <Text className="text-2xl ml-4 mb-4">{t( "Add-evidence" )}</Text>
-            {disableAddingMoreEvidence
-              && (
-              <Text style={textStyles.evidenceWarning}>
-                {t( "You-can-only-upload-20-media" )}
-              </Text>
-              )}
-            <View className="flex-row w-full justify-around">
-              <EvidenceButton
-                icon="perm-media"
-                handlePress={onImportPhoto}
-                disabled={disableAddingMoreEvidence}
-              />
-              <EvidenceButton
-                icon="photo-camera"
-                handlePress={onTakePhoto}
-                disabled={disableAddingMoreEvidence}
-              />
-              <EvidenceButton
-                icon="keyboard-voice"
-                handlePress={onRecordSound}
-                disabled={disableAddingMoreEvidence}
-              />
-            </View>
-            <Text
-              className="underline mt-5"
-              onPress={( () => bottomSheetModalRef.current?.dismiss() )}
-            >
-              {t( "Cancel" )}
-            </Text>
-          </View>
-        </BottomSheetModal>
+        <AddEvidenceModal
+          showAddEvidenceModal={showAddEvidenceModal}
+          setShowAddEvidenceModal={setShowAddEvidenceModal}
+        />
       </ScrollNoFooter>
-    </BottomSheetModalProvider>
+    </>
   );
 };
